+++ Assets/Scripts/Taurus.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Taurus : MonoBehaviour {
 
-    int[,] plan =
+    short[,] plan =
         {
             {0,0,0,0,0,0,1,1,0,0,0,0,0,0,0 },
             {0,1,1,1,1,2,1,1,1,1,0,0,0,0,0 },
